@@ -1,64 +1,25 @@
 {
   "name": "react-router",
-  "version": "5.3.1",
+  "version": "6.3.0",
+  "author": "Remix Software <hello@remix.run>",
   "description": "Declarative routing for React",
-  "homepage": "https://reactrouter.com/",
   "repository": {
-    "url": "https://github.com/remix-run/react-router.git",
     "type": "git",
+    "url": "https://github.com/remix-run/react-router.git",
     "directory": "packages/react-router"
   },
   "license": "MIT",
-  "author": "React Training <hello@reacttraining.com>",
-  "files": [
-    "LICENSE",
-    "README.md",
-    "MemoryRouter.js",
-    "Prompt.js",
-    "Redirect.js",
-    "Route.js",
-    "Router.js",
-    "StaticRouter.js",
-    "Switch.js",
-    "cjs",
-    "es",
-    "esm",
-    "index.js",
-    "generatePath.js",
-    "matchPath.js",
-    "modules/*.js",
-    "modules/utils/*.js",
-    "withRouter.js",
-    "warnAboutDeprecatedCJSRequire.js",
-    "umd"
-  ],
-  "main": "index.js",
-  "module": "esm/react-router.js",
-  "sideEffects": false,
-  "scripts": {
-    "build": "rollup -c",
-    "lint": "eslint modules"
-  },
+  "main": "./main.js",
+  "module": "./index.js",
+  "types": "./index.d.ts",
+  "unpkg": "./umd/react-router.production.min.js",
   "peerDependencies": {
-    "react": ">=15"
+    "react": ">=16.8"
   },
   "dependencies": {
-    "@babel/runtime": "^7.12.13",
-    "history": "^4.9.0",
-    "hoist-non-react-statics": "^3.1.0",
-    "loose-envify": "^1.3.1",
-    "mini-create-react-context": "^0.4.0",
-    "path-to-regexp": "^1.7.0",
-    "prop-types": "^15.6.2",
-    "react-is": "^16.6.0",
-    "tiny-invariant": "^1.0.2",
-    "tiny-warning": "^1.0.0"
+    "history": "^5.2.0"
   },
-  "browserify": {
-    "transform": [
-      "loose-envify"
-    ]
-  },
+  "sideEffects": false,
   "keywords": [
     "react",
     "router",
