@@ -14,7 +14,7 @@ const Productos = () => {
             <img src={product.img} alt={product.name} />
             <div>
               <p>
-                {product.name} - ${product.price}
+                {product.name} - {product.price}€
               </p>
             </div>
             {!product.inCart ? (
